+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-dao/hsweb-system-dynamic-form-dao-api/src/main/java/org/hswebframework/web/dao/form/DynamicFormDao.java
@@ -1,0 +1,11 @@
+package org.hswebframework.web.dao.form;
+
+import org.hswebframework.web.dao.CrudDao;
+import org.hswebframework.web.entity.form.DynamicFormEntity;
+
+/**
+*  动态表单 DAO接口
+*  @author hsweb-generator-online
+ */
+public interface DynamicFormDao extends CrudDao<DynamicFormEntity,String> {
+}
