+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-dao/hsweb-system-dynamic-form-dao-api/src/main/java/org/hswebframework/web/dao/form/DynamicFormDeployLogDao.java
@@ -1,0 +1,11 @@
+package org.hswebframework.web.dao.form;
+
+import org.hswebframework.web.dao.CrudDao;
+import org.hswebframework.web.entity.form.DynamicFormDeployLogEntity;
+
+/**
+*  表单发布日志 DAO接口
+*  @author hsweb-generator-online
+ */
+public interface DynamicFormDeployLogDao extends CrudDao<DynamicFormDeployLogEntity,String> {
+}
