+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-service/hsweb-system-dynamic-form-service-api/src/main/java/org/hswebframework/web/service/form/DatabaseRepository.java
@@ -1,0 +1,18 @@
+package org.hswebframework.web.service.form;
+
+import org.hsweb.ezorm.rdb.RDBDatabase;
+
+/**
+ * 数据库仓库,用于获取默认数据库信息,指定数据源的信息以及当前激活的数据库信息
+ *
+ * @author zhouhao
+ * @see RDBDatabase
+ * @since 3.0
+ */
+public interface DatabaseRepository {
+    RDBDatabase getDefaultDatabase();
+
+    RDBDatabase getDatabase(String datasourceId);
+
+    RDBDatabase getActiveDatabase();
+}
