+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-service/hsweb-system-dynamic-form-service-api/src/main/java/org/hswebframework/web/service/form/DynamicFormColumnService.java
@@ -1,0 +1,13 @@
+package org.hswebframework.web.service.form;
+
+import org.hswebframework.web.entity.form.DynamicFormColumnEntity;
+import org.hswebframework.web.service.CrudService;
+
+/**
+ *  动态表单 服务类
+ *
+ * @author hsweb-generator-online
+ */
+public interface DynamicFormColumnService extends CrudService<DynamicFormColumnEntity, String> {
+
+}
