+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-service/hsweb-system-dynamic-form-service-api/src/main/java/org/hswebframework/web/service/form/DynamicFormOperationService.java
@@ -1,0 +1,24 @@
+package org.hswebframework.web.service.form;
+
+import org.hswebframework.web.commons.entity.PagerResult;
+import org.hswebframework.web.commons.entity.param.DeleteParamEntity;
+import org.hswebframework.web.commons.entity.param.QueryParamEntity;
+import org.hswebframework.web.commons.entity.param.UpdateParamEntity;
+
+/**
+ * TODO 完成注释
+ *
+ * @author zhouhao
+ */
+public interface DynamicFormOperationService {
+    <T> PagerResult<T> selectPager(String formId, QueryParamEntity paramEntity);
+
+    <T> T selectSingle(String formId, QueryParamEntity paramEntity);
+
+    int count(String formId, QueryParamEntity paramEntity);
+
+    <T> int update(String formId, UpdateParamEntity<T> paramEntity);
+
+    int delete(String formId, DeleteParamEntity paramEntity);
+
+}
