+++ hsweb-system/hsweb-system-dynamic-form/hsweb-system-dynamic-form-service/hsweb-system-dynamic-form-service-api/src/main/java/org/hswebframework/web/service/form/OptionalConvertBuilder.java
@@ -1,0 +1,12 @@
+package org.hswebframework.web.service.form;
+
+import org.hsweb.ezorm.core.OptionConverter;
+
+/**
+ * TODO 完成注释
+ *
+ * @author zhouhao
+ */
+public interface OptionalConvertBuilder {
+    OptionConverter buildFromDict(String dictId, String dictParserId);
+}
